Refactor: Move image generation to API route
